% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateDensity.R
\name{calculateDensity}
\alias{calculateDensity}
\title{NIMBLE Function to calculate the density of individuals alive in each habitat cell.}
\usage{
calculateDensity(s, habitatGrid, indicator, numWindows, nIndividuals)
}
\arguments{
\item{s}{\code{Vector} of x- and y-coordinates of individual AC location.}

\item{habitatGrid}{\code{Matrix} Matrix of habitat window indices.}

\item{indicator}{\code{Vector} specifying whether (1) or not (0) individuals are considered alive.}

\item{numWindows}{\code{Scalar} Number of habitat windows.}

\item{nIndividuals}{\code{Scalar} Total number of individuals .}
}
\description{
\code{calculateDensity} is a NIMBLE Function to calculate number of individual activity centers (s) in each habitat cell.
}
\examples{
lowerCoords <- matrix(c(0, 0, 1, 0, 0, 1, 1, 1), nrow = 4, byrow = TRUE)
upperCoords <- matrix(c(1, 1, 2, 1, 1, 2, 2, 2), nrow = 4, byrow = TRUE)  
logIntensities <- log(rep(1,4))
logSumIntensity <- log(sum(c(1:4))) 
habitatGrid <- matrix(c(1:4), nrow = 2, byrow = TRUE)
numGridRows <- nrow(habitatGrid)
numGridCols <- ncol(habitatGrid)

s <- matrix(NA,nrow=10,ncol=2)
for(i in 1:10){
  s[i,] <- rbernppAC(n=1, lowerCoords, upperCoords, logIntensities, logSumIntensity, 
                     habitatGrid, numGridRows, numGridCols)
}

calculateDensity(s = s,
                 habitatGrid = habitatGrid,
                 indicator = rep(1, 10),
                 numWindows = prod(dim(habitatGrid)),
                 nIndividuals = 10
)
}
\author{
Cyril Milleret
}
