% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbernppAC.R
\name{dbernppAC}
\alias{dbernppAC}
\alias{rbernppAC}
\title{Bernoulli point process for the distribution of activity centers}
\usage{
dbernppAC(
  x,
  lowerCoords,
  upperCoords,
  logIntensities,
  logSumIntensity,
  habitatGrid,
  numGridRows,
  numGridCols,
  log = 0
)

rbernppAC(
  n,
  lowerCoords,
  upperCoords,
  logIntensities,
  logSumIntensity,
  habitatGrid,
  numGridRows,
  numGridCols
)
}
\arguments{
\item{x}{Vector of x- and y-coordinates of a single spatial point (i.e. AC location).}

\item{lowerCoords, upperCoords}{Matrices of lower and upper x- and y-coordinates of all habitat windows. One row for each window. 
Each window should be of size 1x1 (after rescaling if necessary).}

\item{logIntensities}{Vector of log habitat intensities for all habitat windows.}

\item{logSumIntensity}{Log of the sum of habitat intensities over all windows.}

\item{habitatGrid}{Matrix of habitat window indices. Only needed for \code{dbernppAC}. 
Habitat window indices should match the order in \code{lowerCoords}, \code{upperCoords}, and \code{logIntensities}. 
When the grid has only one row/column, artificial indices have to be provided to inflate \code{habitatGrid} in order 
to be able to use the distribution in \code{nimble} model code.}

\item{numGridRows, numGridCols}{Numbers of rows and columns of the habitat grid.}

\item{log}{Logical argument, specifying whether to return the log-probability of the distribution.}

\item{n}{Integer specifying the number of realisations to generate.  Only n = 1 is supported.}
}
\value{
\code{dbernppAC} gives the (log) probability density of the observation vector \code{x}. 
\code{rbernppAC} gives coordinates of a randomly generated spatial point.
}
\description{
Density and random generation functions of the Bernoulli point process for the distribution of activity centers.
}
\examples{
# Use the distribution in R
lowerCoords <- matrix(c(0, 0, 1, 0, 0, 1, 1, 1), nrow = 4, byrow = TRUE)
upperCoords <- matrix(c(1, 1, 2, 1, 1, 2, 2, 2), nrow = 4, byrow = TRUE)  
logIntensities <- log(c(1:4))
logSumIntensity <- log(sum(c(1:4)))  
habitatGrid <- matrix(c(1:4), nrow = 2, byrow = TRUE)
numGridRows <- nrow(habitatGrid)
numGridCols <- ncol(habitatGrid)
dbernppAC(c(0.5, 1.5), lowerCoords, upperCoords, logIntensities, logSumIntensity, 
          habitatGrid, numGridRows, numGridCols, log = TRUE)
}
\references{
W. Zhang, J. D. Chipperfield, J. B. Illian, P. Dupont, C. Milleret, P. de Valpine and R. Bischof. 2020. 
A hierarchical point process model for spatial capture-recapture data. bioRxiv. DOI 10.1101/2020.10.06.325035
}
\author{
Wei Zhang
}
