% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcatState2Alive2Dead.R
\name{dcatState2Alive2Dead}
\alias{dcatState2Alive2Dead}
\alias{rcatState2Alive2Dead}
\title{Density and random generation of a categorical distribution describing state transition with two alive and two dead states.}
\usage{
dcatState2Alive2Dead(
  x,
  z,
  prob1To2 = -999,
  prob1To2Hab,
  prob2To3 = -999,
  prob2To3Hab,
  prob2To4 = -999,
  prob2To4Hab,
  prob3To4 = -999,
  prob3To4Hab,
  prob2To5 = -999,
  prob2To5Hab,
  prob3To5 = -999,
  prob3To5Hab,
  s,
  habitatGrid,
  log = 0
)

rcatState2Alive2Dead(
  n,
  z,
  prob1To2 = -999,
  prob1To2Hab,
  prob2To3 = -999,
  prob2To3Hab,
  prob2To4 = -999,
  prob2To4Hab,
  prob3To4 = -999,
  prob3To4Hab,
  prob2To5 = -999,
  prob2To5Hab,
  prob3To5 = -999,
  prob3To5Hab,
  s,
  habitatGrid
)
}
\arguments{
\item{x}{Scalar of the individual state z_{i,t+1}.}

\item{z}{Scalar of the initial individual state z_{i,t}}

\item{prob1To2}{Scalar with probability prob1To2 to transition from z_{i,t} = 1 to z_{i,t+1} = 2.}

\item{prob1To2Hab}{Vector with probability prob1To2_r to transition from z_{i,t} =1 to z_{i,t+1} = 2 for each cell r of the habitat. The vector should be of the length of the number of habitat windows in \code{habitatGrid}.}

\item{prob2To3}{Scalar with probability prob2To3 to transition from z_{i,t} = 1 to z_{i,t+1} = 2.}

\item{prob2To3Hab}{Vector with probability h_r to transition from z_{i,t} = 2 to z_{i,t+1} = 3 for each cell r of the habitat. The vector should be of the length of the number of habitat windows in \code{habitatGrid}.}

\item{prob2To4}{Vector with probability w_r to transition from z_{i,t} = 2 to z_{i,t+1} = 4 for each cell r of the habitat. The vector should be of the length of the number of habitat windows in \code{habitatGrid}.}

\item{prob2To4Hab}{Vector with probability w_r to transition from z_{i,t} = 2 to z_{i,t+1} = 4 for each cell r of the habitat. The vector should be of the length of the number of habitat windows in \code{habitatGrid}.}

\item{prob3To4}{Vector with probability w_r to transition from z_{i,t} = 2 to z_{i,t+1} = 4 for each cell r of the habitat. The vector should be of the length of the number of habitat windows in \code{habitatGrid}.}

\item{prob3To4Hab}{Vector with probability w_r to transition from z_{i,t} = 2 to z_{i,t+1} = 4 for each cell r of the habitat. The vector should be of the length of the number of habitat windows in \code{habitatGrid}.}

\item{prob2To5}{Vector with probability w_r to transition from z_{i,t} = 2 to z_{i,t+1} = 4 for each cell r of the habitat. The vector should be of the length of the number of habitat windows in \code{habitatGrid}.}

\item{prob2To5Hab}{Vector with probability w_r to transition from z_{i,t} = 2 to z_{i,t+1} = 4 for each cell r of the habitat. The vector should be of the length of the number of habitat windows in \code{habitatGrid}.}

\item{prob3To5}{Vector with probability w_r to transition from z_{i,t} = 2 to z_{i,t+1} = 4 for each cell r of the habitat. The vector should be of the length of the number of habitat windows in \code{habitatGrid}.}

\item{prob3To5Hab}{Vector with probability w_r to transition from z_{i,t} = 2 to z_{i,t+1} = 4 for each cell r of the habitat. The vector should be of the length of the number of habitat windows in \code{habitatGrid}.}

\item{s}{Vector of x- and y-coordinates corresponding to the AC location of the individual. Used to extract transition probabilities when spatial probabilities are provided.}

\item{habitatGrid}{Matrix of habitat window indices and only used if arguements prob2To3Hab, prob2To4Hab or phiSpatial are used.
Habitat window indices should match the order in \code{phiSpatial}, \code{prob2To3Hab}, or \code{prob2To4Hab}.}

\item{log}{Logical argument, specifying whether to return the log-probability of the distribution.}

\item{n}{Integer specifying the number of realizations to generate.  Only n = 1 is supported.}
}
\value{
\code{dcatState2Alive2Dead} gives the (log) probability density of \code{x}. 
\code{dcatState2Alive2Dead} gives a randomly generated individual states conditional on the initial state \code{z}.
}
\description{
The \code{dcatState2Alive2Dead} distribution is a NIMBLE custom distribution which can be used to model and simulate
individual state transition. Typically, this is used to model transitions from two alive and two dead states. 
If z_{i,t} = 1, individual i can be recruited (transition to state 2) with probability prob1To2_t, so z_{i,t+1} ~ dcat(1- prob1To2_t, prob1To2_t, 0,0 , 0) where prob1To2_t represent the probability of an unborn individual to be recruited.
If z_{i,t} = 2, individual i can die from one cause of mortality (e.g. culling) and transition to z_{i,t+1}=4 with probability prob2To4, or die from another cause with probability prob2To5 z_{i,t+1}=5. 
If the individual does not die (1-(prob2To4+prob2To5)), it can either transition to the second state alive (z_{i,t+1}=3) with probability prob2To3 or remain in the first state alive (z_{i,t+1}=2) with probability (1-prob2To3). 
If z_{i,t} = 3, individual i can die from one cause of mortality (e.g. culling) and transition to z_{i,t+1}=4 with probability prob3To4, or die from another cause with probability prob3To5 z_{i,t+1}=5. 
if the individual does not die (1-(prob3To4+prob3To5)), the individual remain in state 3. 
Individuals in dead states (z_{i,t} = 4 or 5) transition to z_{i,t+1} = 5, the absorbing state, with probability 1.
If transition probabilities are assumed to be spatially heterogeneous, a vector of probability should be provided using the "Hab" arguments (e.g. prob1To2Hab,prob2To3Hab,..).
}
\examples{
# Use the distribution in R

z <- 3
prob1To2 <- 0.2
prob2To3 <- 0.4
prob2To4 <- 0.1
prob2To5 <- 0.1

prob3To4 <- 0.2
prob3To5 <- 0.1


lowerCoords <- matrix(c(0, 0, 1, 0, 0, 1, 1, 1), nrow = 4, byrow = TRUE)
upperCoords <- matrix(c(1, 1, 2, 1, 1, 2, 2, 2), nrow = 4, byrow = TRUE)  
logIntensities <- log(rep(1,4))
logSumIntensity <- log(sum(c(1:4))) 
habitatGrid <- matrix(c(1:4), nrow = 2, byrow = TRUE)
numGridRows <- nrow(habitatGrid)
numGridCols <- ncol(habitatGrid)
s <- rbernppAC(n=1, lowerCoords, upperCoords, logIntensities, logSumIntensity, 
              habitatGrid, numGridRows, numGridCols)

## No spatial mortality 
zPlusOne <- rcatState2Alive2Dead( z = z
                                 , prob1To2 = prob1To2
                                 , prob2To3 = prob2To3
                                 , prob2To4 = prob2To4
                                 , prob2To5 = prob2To5
                                 , prob3To4 = prob3To4
                                 , prob3To5 = prob3To5
                                 
                                 , s = s
                                 , habitatGrid = habitatGrid)

dcatState2Alive2Dead(  x = zPlusOne
                      , z = z
                      , prob1To2 = prob1To2
                      , prob2To3 = prob2To3
                      , prob2To4 = prob2To4
                      , prob2To5 = prob2To5
                      , prob3To4 = prob3To4
                      , prob3To5 = prob3To5
                      , s = s
                      , habitatGrid = habitatGrid)

##  With spatial mortality
prob2To3Hab <- runif(length(habitatGrid),0,0.1)
prob2To4Hab <- runif(length(habitatGrid),0,0.1)
prob2To5Hab <- runif(length(habitatGrid),0,0.1)
prob3To4Hab <- runif(length(habitatGrid),0,0.1)
prob3To5Hab <- runif(length(habitatGrid),0,0.1)



zPlusOne <- rcatState2Alive2Dead( z = z
                                 , prob1To2 = prob1To2
                                 , prob2To3Hab = prob2To3Hab
                                 , prob2To4Hab = prob2To4Hab
                                 , prob2To5Hab = prob2To5Hab
                                 , prob3To4Hab = prob3To4Hab
                                 , prob3To5Hab = prob3To5Hab
                                 , s = s
                                 , habitatGrid = habitatGrid)

dcatState2Alive2Dead(  x = zPlusOne
                      , z = z
                      , prob1To2 = prob1To2
                      , prob2To3Hab = prob2To3Hab
                      , prob2To4Hab = prob2To4Hab
                      , prob2To5Hab = prob2To5Hab
                      , prob3To4Hab = prob3To4Hab
                      , prob3To5Hab = prob3To5Hab
                      , s = s
                      , habitatGrid = habitatGrid)


}
\author{
Cyril Milleret
}
