% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codeFile.R
\name{dDispersal}
\alias{dDispersal}
\alias{rDispersal}
\title{Bivariate dispersal distribution for activity centers}
\usage{
dDispersal(x, S, lam, log)

rDispersal(n, S, lam)
}
\arguments{
\item{x}{DESCRIPTION OF x PARAMETER}

\item{S}{DESCRIPTION OF S PARAMETER}

\item{lam}{DESCRIPTION OF lam PARAMETER}

\item{log}{DESCRIPTION OF log PARAMETER}

\item{n}{DESCRIPTION OF n PARAMETER}
}
\value{
returns the log-likelihood value associated with the bivariate activity center location x, given the current activity center S, and the lambda parameter of the exponential dispersal distance distribution
}
\description{
Longer description here.  Describe how to use it, etc.
}
\examples{
## R CODE EXAMPLE HERE USING dDispersal DISTRIBUTION
x[1:2] ~ dDispersal()
### etc ...

}
\author{
Daniel Turek
}
