% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exampleFunction.R
\name{exampleFunction}
\alias{exampleFunction}
\title{One phrase description of function.}
\usage{
exampleFunction(a, b)
}
\arguments{
\item{a}{Description of first argument here}

\item{b}{Description of second argument here}
}
\value{
Returns the sum of first two arguments
}
\description{
\code{exampleFunction} returns the sum of the two arguments.  This paragraph, and subsequent paragraphs, contain the longer description of the function.
}
\examples{
x <- exampleFunction(1, 2)

}
