% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HRA_nimble.R
\name{HRA_nimble}
\alias{HRA_nimble}
\title{Function to compute home range radius and area for a specified circular quantile with respect to circular detection functions available in nimbleSCR, viz., half-normal(HN, detfun = 0), half-normal plateau (HNP, detfun = 1), exponential (EX, detfun = 2), asymmetric logistic (AL, detfun = 3), bimodal (BI, detfun = 4) and donut (DN, detfun = 5).}
\usage{
HRA_nimble(
  param = 2,
  xlim = c(0, 30),
  ylim = c(0, 30),
  ng = 800,
  tol = 1e-05,
  niter = 2000,
  detfun = 0,
  pr = 0.95,
  d = 6
)
}
\arguments{
\item{param}{A \code{Vector} of values for different parameters in the specified detection function.}

\item{xlim}{\code{Vector} of length 2 giving the range along x-axis.}

\item{ylim}{\code{Vector} of length 2 giving the range along y-axis.}

\item{ng}{\code{Numeric} variable denoting the number of breaks along an axis.}

\item{tol}{A \code{Numeric} variable denoting the allowed tolerance in the radius estimate.}

\item{niter}{A \code{Numeric} variable giving the maximum number of iterations in bisection algorithm.}

\item{detfun}{\code{Numeric} variable denoting the type of detection function. 0 = Half-normal (HN), 1 = Half-normal plateau (HNP), 2 = Exponential (EX), 3 = Aysmmetric logistic (AL), 4 = Bimodal (BI), 5 = Donut (DN).}

\item{pr}{A \code{Numeric}  variable denoting the quantile probability to compute the radius.}

\item{d}{A \code{Numeric} variable giving an inital value of the radius.}
}
\description{
\code{HRA_nimble} returns the estimates of home range radius and area for a given set of parameters with respect to a particular detection function using bisection algorithm.
}
\author{
Soumen Dey
}
