% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampler_categorical_general.R
\name{sampler_categorical_general}
\alias{sampler_categorical_general}
\title{NIMBLE function to calculate the density of individuals alive in each habitat cell.}
\usage{
sampler_categorical_general(model, mvSaved, target, control)
}
\arguments{
\item{model}{(uncompiled) model on which the MCMC is to be run.}

\item{mvSaved}{\code{modelValues} object to be used to store MCMC samples.}

\item{target}{node(s) on which the sampler will be used.}

\item{control}{named list that controls the precise behavior of the sampler. the number of categories can be provide list("numCategories"= 4)}
}
\description{
\code{sampler_categorical_general} The categorical sampler performs Gibbs sampling for a single node, which must follow a categorical (dcat) distribution.
}
\author{
Daniel Turek
}
