% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaleCoordsToHabitatGrid.R
\name{scaleCoordsToHabitatGrid}
\alias{scaleCoordsToHabitatGrid}
\title{Scale x and y coordinates to the habitat grid}
\usage{
scaleCoordsToHabitatGrid(
  coordsData = coordsData,
  coordsHabitatGridCenter = coordsHabitatGridCenter,
  scaleToGrid = TRUE
)
}
\arguments{
\item{coordsData}{A matrix or an array of the x and y data to scale to the habitat grid. Is is necessary to provide the dimnames of the x and y data such as c("x", "y")}

\item{coordsHabitatGridCenter}{A matrix with the x and y coordinates of each habitat  grid cell center.}

\item{scaleToGrid}{If FALSE, the coordsData are already scaled and will be rescaled to its original coordinates.}
}
\value{
This function returns a list of objects:
\itemize{
\item coordsDataScaled: A matrix or an array of the scaled (rescaled if scaleToGrid==FALSE) x and y coordinates coordsData
\item coordsHabitatGridCenterScaled: A matrix of the scaled x and y cell coordinate centers coordsHabitatGridCenter
}
}
\description{
R utility function to scale x and y coordinates to the habitat grid.
Rescaling the coordinates to the habitat allows a fast look up of a matrix of habitat cellID from x and y coordinates. 
This technique was first applied by Mike Meredith in SCR (\href{https://mmeredith.net/blog/1309_SECR_in_JAGS_patchy_habitat.htm}).
Rescaling the entire coordinate system of the data input is a requirement to run SCR models with the (\code{\link{nimbleSCR}}) package.  
This square grid cells and utm projection (units in meters or km; not latlong!)
}
\details{
Using \code{scaleCoordsToHabitatGrid} to rescale the entire coordinate system of the data input is a requirement to run SCR models with the (\code{\link{nimbleSCR}}) package.
}
\examples{
coordsGridCenter <- expand.grid(list(x = seq(50.5, 100, 1),
                                    y = seq(100.5, 150, 1)))
coordsData <- expand.grid(list(x = seq(60, 90, 1),
                             y = seq(110, 140, 1)))
plot(coordsGridCenter[,1]~coordsGridCenter[,2])
points(coordsData[,1]~coordsData[,2], col="red")
scaled <- scaleCoordsToHabitatGrid(coordsData = coordsData
                                  , coordsHabitatGridCenter = coordsGridCenter)
plot(scaled$coordsHabitatGridCenterScaled[,1]~scaled$coordsHabitatGridCenterScaled[,2])
points(scaled$coordsDataScaled[,1]~scaled$coordsDataScaled[,2], col="red")

}
\author{
Richard Bischof, Cyril Milleret
}
