% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbinom_vector.R
\name{dbinom_vector}
\alias{dbinom_vector}
\alias{rbinom_vector}
\title{Vectorized binomial distribution}
\usage{
dbinom_vector(x, size, prob, log = 0)

rbinom_vector(n = 1, size, prob)
}
\arguments{
\item{x}{Vector of quantiles.}

\item{size}{Vector of number of trials (zero or more).}

\item{prob}{Vector of success probabilities on each trial}

\item{log}{Logical argument, specifying whether to return the log-probability of the distribution.}

\item{n}{Number of observations. Only n = 1 is supported.}
}
\value{
The log-likelihood value associated with the vector of binomial observations.
}
\description{
The \code{dbinom_vector} distribution is a vectorized version of the binomial distribution.
It can be used to model a vector of binomial realizations. NB: using the vectorized version 
is beneficial only when the entire joint likelihood of the vector of binomial realizations (x)
can be calculated simultaneously.
}
\examples{
\donttest{
# simulate binomial data
p <- rep(0.21, 1000)
trials <- sample(x = 10, size = 1000, replace = T)
y <- rbinom_vector(n = 1000, size = trials, prob = p)

## define model code
code <- nimbleCode({
    p ~ dunif(0,1)
    for(j in 1:J) {
      p_vector[i] <- p
      y[j] ~ dbinom( size = trials[i],
                     prob = p_vector[i])
    }
})

## define vectorized model code
code <- nimbleCode({
    p ~ dunif(0,1)
    p_vector[1:J] <- p
    y[1:J] ~ dbinom_vector( size = trials[1:J],
                            prob = p_vector[1:J])
})

## create NIMBLE model object
Rmodel <- nimbleModel(code,data=list(y=y,trials=trials),constants = list(J=length(y)))

## use model object for MCMC, etc.
}

}
\author{
Pierre Dupont
}
