% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSparseY.R
\name{getSparseY}
\alias{getSparseY}
\title{Sparse Matrix Preparation}
\usage{
getSparseY(x, noDetections = -1)
}
\arguments{
\item{x}{a two- or three-dimensional observation data array.}

\item{noDetections}{the default value indicating no detection.}
}
\value{
a list of object which constitute a sparse matrix representation of the observation data provided as input:
- the maximum number of detections for any individual (maxDetNums)
- a vector of the number of detections per individual (detNums).
- a matrix of dimensions n.individuals * maxDetNums which contains the IDs of the detectors 
where each individuals was detected (detIndices).
- a matrix of dimensions n.individuals * maxDetNums which contains the number of observations of each individual
at the detectors it was detected (y).
}
\description{
R utility function to turn a two or three-dimensional detection array into a sparse matrix representation.
}
\details{
The \code{getSparseY} function is used in advance of model building.
It creates and returns a list of objects:
- the maximum number of detections for any individual in the dataset (maxDetNums)
- a vector of the number of detections per individual (detNums).
- a matrix of dimensions n.individuals * maxDetNums which contains the IDs of the detectors 
where each individuals was detected (detIndices).
- a matrix of dimensions n.individuals * maxDetNums which contains the number of observations of each individual
at the detectors it was detected (y).
}
\examples{
\donttest{
y.full <- matrix(rbinom(5000, 5, 0.02), ncol = 100)

y <- getSparseY(y.full)
}

}
\author{
Cyril Milleret
}
